Finished making a test atonomus, set up a 2 motor lift, and removed excess code that is no longer needed
Compiled so it should work

Took our the PNEUMATIC code, will have to pull from earlier verions if we need it back

Note: This autonomous code is UNTESTED and was created prior to robot construction.
Make sure the code is up to  date by using 'git pull'
This should go forward, grab the ring holder, and retreat.
Variables to be changed